LayoutTests/fast/canvas/canvas-scale-strokePath-shadow-expected.txt

Ensure correct behavior of canvas with path stroke + shadow after scaling. A blue and red checkered pattern should be displayed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 76
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 76
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 76
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 20
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 22
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 28
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 22
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 15
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 17
PASS successfullyParsed is true

TEST COMPLETE
